requirements.txt
streamlit
cryptography
azure-identity 
azure-keyvault-secrets
azure-keyvault-keys